Setting the type back to 'jar' from 'war'.
The current version is not built as a war file, but rather as a
standalone jar which deploys using an embedded Jetty server.

This isn't the way it should be done, so one of the next major updates
will be moving this to be built and deployed as a war file directly.