Load JSON schema types from classpath resources.
java-json-schema v1.0.1 now supports loading schema types from the
classpath as resources, using the JSONResourceType class.

SkeletonServlet had been using the File constructor to SchemaEnv, which
loads the types from the filesystem ; however, this configuration won't
work in an environment where we're deploying as a WAR file.

So instead, I've converted SkeletonServlet to load all the types it
expects from the classpath as resources.  All the schemas are currently
in the package 'schemas'.